+++ src/utilities/testing/mocking/mockSubscriptionLink.ts
@@ -17,7 +17,7 @@ export class MockSubscriptionLink extends ApolloLink {
   public setups: any[] = [];
   public operation: Operation;
 
-  private observer: any;
+  private observers: any[] = [];
 
   constructor() {
     super();
@@ -27,7 +27,7 @@ export class MockSubscriptionLink extends ApolloLink {
     this.operation = operation;
     return new Observable<FetchResult>(observer => {
       this.setups.forEach(x => x());
-      this.observer = observer;
+      this.observers.push(observer);
       return () => {
         this.unsubscribers.forEach(x => x());
       };
@@ -36,18 +36,22 @@ export class MockSubscriptionLink extends ApolloLink {
 
   public simulateResult(result: MockedSubscriptionResult, complete = false) {
     setTimeout(() => {
-      const { observer } = this;
-      if (!observer) throw new Error('subscription torn down');
-      if (complete && observer.complete) observer.complete();
-      if (result.result && observer.next) observer.next(result.result);
-      if (result.error && observer.error) observer.error(result.error);
+      const { observers } = this;
+      if (!observers.length) throw new Error('subscription torn down');
+      observers.forEach(observer => {
+        if (complete && observer.complete) observer.complete();
+        if (result.result && observer.next) observer.next(result.result);
+        if (result.error && observer.error) observer.error(result.error);
+      });
     }, result.delay || 0);
   }
 
   public simulateComplete() {
-    const { observer } = this;
-    if (!observer) throw new Error('subscription torn down');
-    if (observer.complete) observer.complete();
+    const { observers } = this;
+    if (!observers.length) throw new Error('subscription torn down');
+    observers.forEach(observer => {
+      if (observer.complete) observer.complete();
+    })
   }
 
   public onSetup(listener: any): void {